// main/src/app/shared/models/dtos/UserUpdateDto.ts
// creator: Nguyễn Xuân Hùng
export interface UserUpdateDto {
  'id': number;
  'fullName': string;
  'email': string;
  'birthday': string;
  'password': string;
  'newPassword': string;
  'confirmPassword': string;
  'idCard': string;
  'phoneNumber': string;
  'gender': string;
  'address': string;
  'backendMessage': string[];
}